fix OpenAPI schema warnings
Only tell ChatGPT about the API routes that it can use. We were adding
/.well-known/ai-plugin.json and /s/:id to the OpenAPI schema by using
the default router.get methods which says this route should be included
in the OpenAPI spec.

Use router.original.get for the url shortening service, the homepage,
and other utility routes which avoids adding the route to the OpenAPI
spec.